# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BE API — WP-02 Identidad y sesiones · WP-03 Vínculo, consentimiento y PDP",
    "version": "0.3.0",
    "description": "Generado desde @be/domain (contratos.ts y contratos-vinculo.ts). No editar a mano."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "components": {
    "securitySchemes": {
      "sesion": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Sesión revocable verificada en cada request (08 §26)."
      }
    },
    "schemas": {
      "ErrorEnvelope": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "details": {
                "type": "object",
                "propertyNames": {
                  "type": "string"
                },
                "additionalProperties": {}
              }
            },
            "required": [
              "code",
              "message"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      }
    }
  },
  "paths": {
...
    "/advisees/{adviseeId}/dashboard": {
      "get": {
        "operationId": "API-DSH-03",
        "summary": "Dashboard del asesorado para un profesional, evaluado por el PDP en cada alcance. Sin datos de dominio en WP-03. Ningún alcance autorizado → 404 idéntico a inexistente.",
        "description": "Fuente: 09v11:895-950 · 08:581, 09:715 (SESSION con datos sintéticos) · DEUDA_LEGAJO DL-031. AuthN: SESSION.",
        "security": [
          {
            "sesion": []
          }
        ],
        "parameters": [
          {
            "name": "adviseeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "periodStart",
            "in": "query",
            "required": false,
            "description": "Inicio del período (RFC 3339).",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "periodEnd",
            "in": "query",
            "required": false,
            "description": "Fin del período (RFC 3339).",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "X-BE-Surface",
            "in": "header",
            "required": false,
            "description": "Superficie declarada (procedencia). Nunca autoriza (DEUDA_LEGAJO DL-022).",
            "schema": {
              "type": "string",
              "enum": [
                "WEB",
                "APK"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Éxito",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "advisee": {
                          "type": "object",
                          "properties": {
                            "identityId": {
                              "type": "string",
                              "minLength": 1
                            },
                            "displayName": {
                              "type": "string"
                            }
                          },
                          "required": [
                            "identityId",
                            "displayName"
                          ],
                          "additionalProperties": false
                        },
                        "period": {
                          "type": "object",
                          "properties": {
                            "start": {
                              "anyOf": [
                                {
                                  "type": "string",
                                  "format": "date-time",
                                  "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
                                },
                                {
                                  "type": "null"
                                }
                              ]
                            },
                            "end": {
                              "anyOf": [
                                {
                                  "type": "string",
                                  "format": "date-time",
                                  "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
                                },
                                {
                                  "type": "null"
                                }
                              ]
                            }
                          },
                          "required": [
                            "start",
                            "end"
                          ],
                          "additionalProperties": false
                        },
                        "partialView": {
                          "type": "boolean"
                        },
                        "domains": {
                          "type": "object",
                          "properties": {
                            "nutrition": {
                              "anyOf": [
                                {
                                  "type": "object",
                                  "properties": {
                                    "available": {
                                      "type": "boolean",
                                      "const": true
                                    },
                                    "relationshipId": {
                                      "type": "string",
                                      "minLength": 1
                                    },
                                    "summary": {
                                      "type": "null"
                                    }
                                  },
                                  "required": [
                                    "available",
                                    "relationshipId",
                                    "summary"
                                  ],
                                  "additionalProperties": false
                                },
                                {
                                  "type": "object",
                                  "properties": {
                                    "available": {
                                      "type": "boolean",
                                      "const": false
                                    },
                                    "reason": {
                                      "type": "string",
                                      "const": "NOT_AVAILABLE_TO_VIEW"
                                    }
                                  },
                                  "required": [
                                    "available",
                                    "reason"
                                  ],
                                  "additionalProperties": false
                                }
                              ]
                            },
                            "training": {
                              "anyOf": [
                                {
                                  "type": "object",
                                  "properties": {
                                    "available": {
                                      "type": "boolean",
                                      "const": true
                                    },
                                    "relationshipId": {
                                      "type": "string",
                                      "minLength": 1
                                    },
                                    "summary": {
                                      "type": "null"
                                    }
                                  },
                                  "required": [
                                    "available",
                                    "relationshipId",
                                    "summary"
                                  ],
                                  "additionalProperties": false
                                },
                                {
                                  "type": "object",
                                  "properties": {
                                    "available": {
                                      "type": "boolean",
                                      "const": false
                                    },
                                    "reason": {
                                      "type": "string",
                                      "const": "NOT_AVAILABLE_TO_VIEW"
                                    }
                                  },
                                  "required": [
                                    "available",
                                    "reason"
                                  ],
                                  "additionalProperties": false
                                }
                              ]
                            },
                            "anthropometry": {
                              "anyOf": [
                                {
                                  "type": "object",
                                  "properties": {
                                    "available": {
                                      "type": "boolean",
                                      "const": true
                                    },
                                    "relationshipId": {
                                      "type": "string",
                                      "minLength": 1
                                    },
                                    "summary": {
                                      "type": "null"
                                    }
                                  },
                                  "required": [
                                    "available",
                                    "relationshipId",
                                    "summary"
                                  ],
                                  "additionalProperties": false
                                },
                                {
                                  "type": "object",
                                  "properties": {
                                    "available": {
                                      "type": "boolean",
                                      "const": false
                                    },
                                    "reason": {
                                      "type": "string",
                                      "const": "NOT_AVAILABLE_TO_VIEW"
                                    }
                                  },
                                  "required": [
                                    "available",
                                    "reason"
                                  ],
                                  "additionalProperties": false
                                }
                              ]
                            }
                          },
                          "required": [
                            "nutrition",
                            "training",
                            "anthropometry"
                          ],
                          "additionalProperties": false
                        }
                      },
                      "required": [
                        "advisee",
                        "period",
                        "partialView",
                        "domains"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "data"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "ErrorEnvelope: INVALID_REQUEST",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "ErrorEnvelope: AUTHENTICATION_REQUIRED | SESSION_INVALID | SESSION_EXPIRED | SESSION_REVOKED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "ErrorEnvelope: RESOURCE_NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "ErrorEnvelope: INTERNAL_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "ErrorEnvelope: DB_UNAVAILABLE",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    }
  }
}